+++ projects/client/src/components/admin/manage_user/UserMgtUserData.jsx
@@ -2,6 +2,8 @@ import React, { useEffect, useState } from "react";
 import { useDispatch } from "react-redux";
 import { getAllUserData, getSingleUser } from "../../../feature/admin/AdminSlice";
 import SingleUserModal from "./all_user/SingleUserModal";
+import RenderUsers from "./all_user/RenderUsers";
+import Pagination from "./all_user/Pagination";
 
 function UserMgtUserData() {
   const dispatch = useDispatch();
@@ -16,14 +18,14 @@ function UserMgtUserData() {
     setIsLoading(false);
   };
 
+  useEffect(() => {
+    dataRender();
+  }, [pageNum]);
+
   const dataClicked = async (id, isAdmin, idRole) => {
     await dispatch(getSingleUser(id, isAdmin, idRole));
     setIsModalPop(true);
   };
-
-  useEffect(() => {
-    dataRender();
-  }, [pageNum]);
 
   const addPageNum = () => {
     setPageNum(pageNum + 1);
@@ -31,28 +33,6 @@ function UserMgtUserData() {
 
   const minusPageNum = () => {
     setPageNum(pageNum - 1);
-  };
-
-  const cutString = (string) => {
-    return string.length > 12 ? string.slice(0, 10) + "..." : string;
-  };
-
-  const RenderUserData = () => {
-    return allUserData.dataAll.map((data) => {
-      return (
-        <div
-          key={data.id_user}
-          className="row-span-1 bg-slate-100 px-2 grid
-            grid-cols-6 items-center hover:cursor-pointer"
-          onClick={() => dataClicked(data.id_user, data.is_admin, data.id_role)}
-        >
-          <p className="col-span-1">{data.id_user}</p>
-          <p className="col-span-2">{cutString(data.username)}</p>
-          <p className="col-span-2">{cutString(data.email)}</p>
-          <p className="col-span-1 text-right">{data.is_admin ? "admin" : "user"}</p>
-        </div>
-      );
-    });
   };
 
   return (
@@ -68,34 +48,18 @@ function UserMgtUserData() {
               <p className="col-span-1 text-right">role</p>
             </div>
             <div className=" row-span-6 grid grid-rows-8 gap-2">
-              <RenderUserData />
+              <RenderUsers allUserData={allUserData} dataClicked={dataClicked} />
             </div>
             <div
               className="items-center row-span-1 py-2 grid grid-cols-7 text-slate-800
-    text-lg"
+              text-lg"
             >
-              <div
-                className="col-span-1 col-start-3 flex items-center 
-      justify-center"
-              >
-                <button onClick={minusPageNum} disabled={pageNum === 1}>
-                  <i className="uil uil-arrow-left hover:cursor-pointer"></i>
-                </button>
-              </div>
-              <div
-                className="col-span-1 col-start-4 flex items-center 
-      justify-center"
-              >
-                <p>{pageNum}</p>
-              </div>
-              <div
-                className="col-span-1 col-start-5 flex items-center 
-      justify-center"
-              >
-                <button onClick={addPageNum} disabled={pageNum === allUserData.totalPage}>
-                  <i className="uil uil-arrow-right hover:cursor-pointer"></i>
-                </button>
-              </div>
+              <Pagination
+                allUserData={allUserData}
+                addPageNum={addPageNum}
+                minusPageNum={minusPageNum}
+                pageNum={pageNum}
+              />
             </div>
           </div>
         </>
