+++ projects/client/src/components/admin/manage_user/all_user/SingleUserModal.jsx
@@ -1,5 +1,7 @@
 import React, { useState } from "react";
 import { useSelector } from "react-redux";
+import DetailUser from "./DetailUser";
+import HeaderModal from "./HeaderModal";
 
 function SingleUserModal(props) {
   const singleUser = useSelector((state) => state.admin.singleUser);
@@ -23,66 +25,12 @@ function SingleUserModal(props) {
         >
           <i className="uil uil-times-circle"></i>
         </button>
-        <div className="">
+        <div>
           <div className="flex justify-between items-center">
-            <h1 className="font-semibold">Detail Data ID : {singleUser.id_user}</h1>
-            <div>
-              {singleUser.role !== "super-admin" ? (
-                <button className="bg-slate-200 text-slate-800 py-1 px-2" onClick={() => setEditState(true)}>
-                  <i class="uil uil-pen"></i> edit
-                </button>
-              ) : null}
-            </div>
+            <HeaderModal singleUser={singleUser} setEditState={setEditState} />
           </div>
           <div className="mt-10 mb-4 flex flex-col gap-2">
-            <div className="flex flex-rows gap-2 text-sm">
-              <div>username : </div>
-              <div>{singleUser.username}</div>
-            </div>
-            <div className="flex flex-rows gap-2 text-sm">
-              <div>email : </div>
-              <div>{singleUser.email}</div>
-            </div>
-            <div className="flex flex-rows gap-2 text-sm">
-              <div>phone number : </div>
-              <div>{singleUser.phone_number}</div>
-            </div>
-            <div className="flex flex-rows gap-2 text-sm">
-              <div>role : </div>
-              <div>{singleUser.role === "user" ? "non-admin" : `${singleUser.role}`}</div>
-              <div>
-                {singleUser.role === "user" ? (
-                  <button
-                    className="px-2 py-1 text-xs bg-slate-700 text-white disabled:bg-slate-100 disabled:text-slate-400"
-                    disabled={!editState}
-                  >
-                    make admin
-                  </button>
-                ) : (
-                  <>
-                    {singleUser.role !== "super-admin" ? (
-                      <button
-                        className="px-2 py-1 text-xs bg-slate-700 text-white disabled:bg-slate-100 
-                        disabled:text-slate-400 disabled:cursor-auto"
-                        disabled={!editState}
-                        onClick={() => setIsAdminClicked(true)}
-                      >
-                        revoke admin
-                      </button>
-                    ) : null}
-                  </>
-                )}
-              </div>
-            </div>
-            {singleUser.role === "admin-warehouse" ? (
-              <div className="flex flex-rows gap-2 text-sm">
-                <div>warehouse : </div>
-                <div>
-                  {singleUser.warehouse}, {singleUser.city}
-                </div>
-                <div></div>
-              </div>
-            ) : null}
+            <DetailUser singleUser={singleUser} editState={editState} setIsAdminClicked={setIsAdminClicked} />
           </div>
           {isAdminClicked ? (
             <div className="flex flex-row gap-4">
